Added Hugh's recommendations

+++ todo_app/app.py
@@ -24,7 +24,6 @@ def load_user(github_user):
 
 def create_app():
     app = Flask(__name__)
-    app.config.from_object(Config)
     app.secret_key = os.environ.get('SECRET_KEY')
     app.config['LOGIN_DISABLED'] = os.environ.get('LOGIN_DISABLED', 'False').lower() in ['true', '1']
     login_manager.init_app(app)
@@ -36,7 +35,7 @@ def create_app():
         items = get_todo_cards()
         item_view_model = ViewModel(items, alllistids)
 
-        if hasattr('current_user', 'role'):
+        if not app.config['LOGIN_DISABLED']:
             if current_user.role == 'writer':
                 return render_template('index.html', view_model=item_view_model)
             elif current_user.role == 'reader':
@@ -44,43 +43,46 @@ def create_app():
             else:
                 return render_template('indexreadonly.html', view_model=item_view_model)
         else:
-            return render_template('indexreadonly.html', view_model=item_view_model)
+            return render_template('index.html', view_model=item_view_model)
 
 
     @app.route('/add_new', methods=['POST'])
     @login_required
     def addtolist():
-        newitemtitle = request.form.get('title')
-        newitemdescription = request.form.get('desc')
-        create_todo_card(newitemtitle, newitemdescription)
+        if 'LOGIN_DISABLED' in app.config or current_user.role == 'writer':
+            newitemtitle = request.form.get('title')
+            newitemdescription = request.form.get('desc')
+            create_todo_card(newitemtitle, newitemdescription)
         return redirect(request.headers.get('Referer'))
 
     @app.route('/update', methods=['POST'])
     @login_required
     def movetolist():
-        allcards = get_todo_cards()
-        lists = {'ToDo':'ToDo','Doing':'Doing','Done':'Done'}
-        for card in allcards:
-            if request.form.get("inprogcheck_" + str(card.id)) == str(card.id):
-                for l in lists:
-                    if l == "Doing":
-                        desired_list_id = l
-                        move_todo_card(card.id, desired_list_id)
-            if request.form.get("donecheck_" + str(card.id)) == str(card.id):
-                for l in lists:
-                    if l == "Done":
-                        desired_list_id = l
-                        move_todo_card(card.id, desired_list_id)
-            if request.form.get("deletecheck_" + str(card.id)) == str(card.id):
-                delete_todo_card(card.id)
+        if 'LOGIN_DISABLED' in app.config or current_user.role == 'writer':
+            allcards = get_todo_cards()
+            lists = {'ToDo':'ToDo','Doing':'Doing','Done':'Done'}
+            for card in allcards:
+                if request.form.get("inprogcheck_" + str(card.id)) == str(card.id):
+                    for l in lists:
+                        if l == "Doing":
+                            desired_list_id = l
+                            move_todo_card(card.id, desired_list_id)
+                if request.form.get("donecheck_" + str(card.id)) == str(card.id):
+                    for l in lists:
+                        if l == "Done":
+                            desired_list_id = l
+                            move_todo_card(card.id, desired_list_id)
+                if request.form.get("deletecheck_" + str(card.id)) == str(card.id):
+                    delete_todo_card(card.id)
         return redirect(request.headers.get('Referer'))
 
     @app.route('/show_older_done_items', methods=['POST'])
     @login_required
     def show_older_completed_items():
-        alllistids = {'ToDo':'ToDo','Doing':'Doing','Done':'Done'}
-        items = get_todo_cards()
-        item_view_model = ViewModel(items, alllistids)
+        if 'LOGIN_DISABLED' in app.config or current_user.role == 'writer':
+            alllistids = {'ToDo':'ToDo','Doing':'Doing','Done':'Done'}
+            items = get_todo_cards()
+            item_view_model = ViewModel(items, alllistids)
         return render_template('showolderitems.html', view_model=item_view_model)
     
     @app.route('/login/callback')
